+++ lib/commands/types.ts
@@ -337,6 +337,60 @@ export type ButtonName = AnyCase<
  */
 export type Style = 'dark' | 'light' | 'unsupported' | 'unknown';
 
+/**
+ * Returned in the {@linkcode XCUITest.mobileGetIncreaseContrast mobile: getIncreaseContrast} command response.
+ */
+export type IncreaseContrastResult =
+    | 'enabled'
+    | 'disabled'
+    | 'unsupported'
+    | 'unknown';
+
+/**
+ * Argument in the {@linkcode XCUITest.mobileSetIncreaseContrast mobile: setIncreaseContrast} command.
+ */
+export type IncreaseContrastAction =
+    | 'enabled'
+    | 'disabled';
+
+/**
+ * Argument in the {@linkcode XCUITest.mobileSetContentSize mobile: setContentSize} command.
+ */
+export type ContentSizeAction =
+    | 'extra-small'
+    | 'small'
+    | 'medium'
+    | 'large'
+    | 'extra-large'
+    | 'extra-extra-large'
+    | 'extra-extra-extra-large'
+    | 'accessibility-medium'
+    | 'accessibility-large'
+    | 'accessibility-extra-large'
+    | 'accessibility-extra-extra-large'
+    | 'accessibility-extra-extra-extra-large'
+    | 'increment'
+    | 'decrement';
+
+/**
+ * Returned in the {@linkcode XCUITest.mobileGetContentSize mobile: getContentSize} command response.
+ */
+export type ContentSizeResult =
+    | 'extra-small'
+    | 'small'
+    | 'medium'
+    | 'large'
+    | 'extra-large'
+    | 'extra-extra-large'
+    | 'extra-extra-extra-large'
+    | 'accessibility-medium'
+    | 'accessibility-large'
+    | 'accessibility-extra-large'
+    | 'accessibility-extra-extra-large'
+    | 'accessibility-extra-extra-extra-large'
+    | 'unknown'
+    | 'unsupported';
+
 export interface ScreenInfo {
   /**
    * Status bar dimensions